+++ README.md
@@ -9,8 +9,8 @@
 
 ### 기능명세서
 
-- [ ] 궁의 이동을 구현한다
-    - [ ] `전`,`후`,`좌`,`우`로 한 칸 이동한다.
+- [x] 궁의 이동을 구현한다
+    - [x] `전`,`후`,`좌`,`우`로 한 칸 이동한다.
 - [ ] 사의 이동을 구현한다
     - [ ] `전`,`후`,`좌`,`우`로 한 칸 이동한다.
 - [ ] 졸의 이동을 구현한다
