--- src/main/java/domain/Piece.java
@@ -1,15 +1,0 @@
-package domain;
-
-import java.util.Map;
-
-public interface Piece {
-    boolean isHan();
-
-    boolean isCho();
-
-    boolean isMovable(Map<Position, Piece> pieces, Position departure, Position destination);
-
-    boolean isSameSide(Piece destinationPiece);
-
-    boolean isPo();
-}
