+++ src/main/java/movepolicy/destination/DestinationRule.java
@@ -1,4 +1,6 @@
-package domain;
+package movepolicy.destination;
+
+import pieces.Piece;
 
 public interface DestinationRule {
     boolean validateDestination(Piece departurePiece, Piece destinationPiece);
