+++ src/main/java/movepolicy/path/PathRule.java
@@ -1,6 +1,7 @@
-package domain;
+package movepolicy.path;
 
 import java.util.List;
+import pieces.Piece;
 
 public interface PathRule {
     boolean validatePathPieces(List<Piece> pathPieces);
