+++ src/main/java/pieces/Piece.java
@@ -1,0 +1,16 @@
+package pieces;
+
+import movepolicy.MoveContext;
+import position.Position;
+
+public interface Piece {
+    boolean isHan();
+
+    boolean isCho();
+
+    MoveContext askMoveContext(Position departure, Position destination);
+
+    boolean isSameSide(Piece destinationPiece);
+
+    boolean isPo();
+}
