+++ src/main/java/position/Column.java
@@ -1,4 +1,4 @@
-package domain;
+package position;
 
 public record Column(int index) {
     private static final int MINIMUM_BOUNDARY = 0;
