+++ src/test/java/movepolicy/path/EmptyPathRuleTest.java
@@ -1,10 +1,13 @@
-package domain;
+package movepolicy.path;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import pieces.Gung;
+import pieces.Piece;
+import pieces.Side;
 
 class EmptyPathRuleTest {
