+++ src/test/java/position/ColumnTest.java
@@ -1,4 +1,4 @@
-package domain;
+package position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,63 +7,62 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-class RowTest {
+class ColumnTest {
     private static final int MINIMUM_BOUNDARY = 0;
-    private static final int MAXIMUM_BOUNDARY = 9;
+    private static final int MAXIMUM_BOUNDARY = 8;
     private static final int ONE_SPACE = 1;
 
     @Test
-    void ROW의_범위가_9를_넘을_경우_예외를_던진다() {
-        assertThatThrownBy(() -> new Row(10))
+    void COLUMN의_범위가_8을_넘을_경우_예외를_던진다() {
+        assertThatThrownBy(() -> new Column(9))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Test
-    void ROW의_범위가_0보다_작을_경우_예외를_던진다() {
-        assertThatThrownBy(() -> new Row(-1))
+    void COLUMN의_범위가_0보다_작을_경우_예외를_던진다() {
+        assertThatThrownBy(() -> new Column(-1))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     @Nested
-    @DisplayName("ROW의 이동을 검증한다")
+    @DisplayName("COLUMN의 이동을 검증한다")
     class Move {
         @Test
-        void ROW가_위로_한_칸_이동한다() {
+        void COLUMN이_오른쪽으로_한_칸_이동한다() {
             // given
-            Row prev = new Row(1);
+            Column prev = new Column(1);
             // when
-            Row cur = prev.up();
+            Column cur = prev.right();
             // then
             assertThat(cur.index()).isEqualTo(prev.index() + ONE_SPACE);
         }
 
         @Test
-        void ROW가_아래로_한_칸_이동한다() {
+        void COLUMN이_왼쪽으로_한_칸_이동한다() {
             // given
-            Row prev = new Row(1);
+            Column prev = new Column(1);
             // when
-            Row cur = prev.down();
+            Column cur = prev.left();
             // then
             assertThat(cur.index()).isEqualTo(prev.index() - ONE_SPACE);
         }
 
         @Test
-        void ROW가_위로_한_칸_이동했을_때_범위를_벗어날경우_예외를_던진다() {
+        void COLUMN이_오른쪽으로_한_칸_이동했을_때_범위를_벗어날경우_예외를_던진다() {
             // given
-            Row prev = new Row(MAXIMUM_BOUNDARY);
+            Column prev = new Column(MAXIMUM_BOUNDARY);
             // when & then
-            assertThatThrownBy(prev::up)
+            assertThatThrownBy(prev::right)
                     .isInstanceOf(IllegalArgumentException.class);
         }
 
         @Test
-        void ROW가_아래로_한_칸_이동했을_때_범위를_벗어날경우_예외를_던진다() {
+        void COLUMN이_왼쪽으로_한_칸_이동했을_때_범위를_벗어날경우_예외를_던진다() {
             // given
-            Row prev = new Row(MINIMUM_BOUNDARY);
+            Column prev = new Column(MINIMUM_BOUNDARY);
             // when & then
-            assertThatThrownBy(prev::down)
+            assertThatThrownBy(prev::left)
                     .isInstanceOf(IllegalArgumentException.class);
         }
     }
-
 }
