+++ src/test/java/position/PositionTest.java
@@ -1,4 +1,4 @@
-package domain;
+package position;
 
 import static org.assertj.core.api.Assertions.assertThat;
